js/xpconnect/tests/mochitest/test_bug384632.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=384632
-->
<head>
  <title>Test for Bug 384632</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=384632">Mozilla Bug 384632</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 384632 **/
var Cc = SpecialPowers.Cc, Ci = SpecialPowers.Ci;
var propBag = Cc["@mozilla.org/hash-property-bag;1"].createInstance(Ci.nsIWritablePropertyBag);
var obj = {};
propBag.setProperty("foopy", obj);
ok(SpecialPowers.unwrap(propBag.getProperty("foopy")) === obj,
   "nsIVariant works with regular objects");
propBag.setProperty("foopy1", external);
ok(SpecialPowers.unwrap(propBag.getProperty("foopy1")) === external,
   "nsIVariant works with bizarre objects");

</script>
</pre>
</body>
</html>
